Moved COvid19CSV to its own file.
