+++ transform/Covid19Utils.py
@@ -1,4 +1,4 @@
-from transform.Covid19CSVTransformer import Covid19CSV
+from transform.Covid19CSV import Covid19CSV
 from data.ObsCollection import ObsCollection
 from data.Location import Location
 from data.ObservationDate import ObservationDate
